refactor code for gpu tile stats + make scheduler stats + clean up

+++ gl_engine/MapLabels.h
@@ -61,7 +61,7 @@ public:
 
     void update_labels(const std::vector<nucleus::vector_tile::PoiTile>& updated_tiles, const std::vector<TileId>& removed_tiles);
 
-    size_t tile_count() const;
+    unsigned int tile_count() const;
 
 private:
     void upload_to_gpu(const TileId& id, const PointOfInterestCollection& features);
